move codegen to isomorphic, generate code on the client side
