+++ packages/playwright-core/src/client/DEPS.list
@@ -1,3 +1,4 @@
 [*]
 ../protocol/
 ../utils/isomorphic
+../utils/isomorphic/codegen
