--- packages/playwright-core/src/server/codegen/DEPS.list
@@ -1,3 +1,0 @@
-[*]
-../../utils/
-../deviceDescriptors.ts
