+++ packages/playwright-core/src/utils/isomorphic/codegen/DEPS.list
@@ -1,0 +1,2 @@
+[*]
+../**
