+++ packages/playwright-core/src/utils/isomorphic/codegen/jsonl.ts
@@ -14,9 +14,9 @@
  * limitations under the License.
  */
 
-import { asLocator } from '../../utils';
+import { asLocator } from '../locatorGenerators';
 
-import type { Language, LanguageGenerator, LanguageGeneratorOptions } from './types';
+import type { Language, LanguageGenerator, LanguageGeneratorOptions } from './codegen';
 import type * as actions from '@recorder/actions';
 
 export class JsonlLanguageGenerator implements LanguageGenerator {
