+++ packages/playwright-core/src/utils/isomorphic/deviceDescriptorsSource.json
