+++ tests/library/browsercontext-viewport.spec.ts
@@ -19,7 +19,7 @@ import { devices } from '@playwright/test';
 import { contextTest as it, expect } from '../config/browserTest';
 import { browserTest } from '../config/browserTest';
 import { verifyViewport } from '../config/utils';
-import { deviceDescriptors } from 'packages/playwright-core/lib/server/deviceDescriptors';
+import { deviceDescriptors } from 'packages/playwright-core/lib/utils/isomorphic/deviceDescriptors';
 
 it('should get the proper default viewport size', async ({ page, server }) => {
   await verifyViewport(page, 1280, 720);
